Src/requirements.txt
fastapi
uvicorn
mangum
openai
pinecone-client
pydantic